e-wallet-back/src/main/java/org/kuehnenagel/config/JwtFilter.java
package org.kuehnenagel.config;

import org.kuehnenagel.service.UserDetailService;
import org.kuehnenagel.util.JwtUtil;
import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
import org.springframework.security.core.context.SecurityContextHolder;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.web.authentication.WebAuthenticationDetailsSource;
import org.springframework.stereotype.Component;
import org.springframework.web.filter.OncePerRequestFilter;

import javax.servlet.FilterChain;
import javax.servlet.ServletException;
import javax.servlet.http.HttpServletRequest;
import javax.servlet.http.HttpServletResponse;
import java.io.IOException;
import java.util.Objects;

@Component
public class JwtFilter extends OncePerRequestFilter {
    
    private final JwtUtil jwtUtil;
    private final UserDetailService userDetailService;
    
    public JwtFilter(JwtUtil jwtUtil, UserDetailService userDetailService) {
        this.jwtUtil = jwtUtil;
        this.userDetailService = userDetailService;
    }
    
    /**
     * Filter any incoming request for valid token
     */
    @Override
    protected void doFilterInternal(
            HttpServletRequest httpServletRequest,
            HttpServletResponse httpServletResponse,
            FilterChain filterChain) throws ServletException, IOException {
        String header = httpServletRequest.getHeader("Authorization");
        
        if (Objects.nonNull(header) && header.startsWith("Bearer")) {
            final String token = header.split(" ")[1];
            final String username = jwtUtil.getUsername(token);
    
            UserDetails userDetails = userDetailService.loadUserByUsername(username);
            var uPAuth = new UsernamePasswordAuthenticationToken(
                    userDetails, null, userDetails.getAuthorities());
            uPAuth.setDetails(new WebAuthenticationDetailsSource().buildDetails(httpServletRequest));
            SecurityContextHolder.getContext().setAuthentication(uPAuth);
        }
        filterChain.doFilter(httpServletRequest, httpServletResponse);
    }
}
